Removed old failure models
