--- opendc-compute/opendc-compute-simulator/src/main/kotlin/org/opendc/compute/simulator/failure/hostfault/HostFaultNew.kt
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2021 AtLarge Research
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package org.opendc.compute.simulator.failure.hostfault
-
-import org.opendc.compute.service.ComputeService
-import org.opendc.compute.simulator.SimHost
-import java.time.InstantSource
-
-/**
- * Interface responsible for applying the fault to a host.
- */
-public abstract class HostFaultNew(
-    private val service: ComputeService,
-    private val clock: InstantSource,
-) {
-    /**
-     * Apply the fault to the specified [victims].
-     */
-    public abstract suspend fun apply(
-        victims: List<SimHost>,
-        faultDuration: Long
-    )
-}
