+++ opendc-compute/opendc-compute-simulator/src/main/kotlin/org/opendc/compute/simulator/failure/prefab/PrefabFailureModelFactory.kt
